debug

+++ examples/mxnet/gc-mc/data.py
@@ -94,7 +94,7 @@ class MovieLens(object):
         rating_pairs = np.array(np.array([self.global_user_id_map[ele]
                                           for ele in rating_info["user_id"]], dtype=np.int64),
                                 np.array([self.global_movie_id_map[ele]
-                                          for ele in rating_info["movie_id"]], dtype=np.int64))
+                                          for ele in rating_info["movie_id"]], dtype=np.int64), dtype=np.int64)
         rating_values = rating_info["rating"].values.astype(np.float32)
         print("rating_pairs", rating_pairs.shape, rating_pairs)
         return rating_pairs, rating_values
